v1.5
Add A star algorithm. Almost done!

+++ AI/src/transport/CarTimetable.java
@@ -38,7 +38,7 @@ public class CarTimetable implements TimeTable  {
         return price;
     }
 
-    public int getDistance() {
+    public double getDistance() {
         return distance;
     }
